Fix issue #60. Replace gitRepositoryRoot config key by dotGitDirectory config key. Improve logic to detect .git directory.  Fix issue with exception thrown when (failOnNoGitDirectory = false and dotGitDirectory = null). Refactor methods to classes to support unit tests later
